fixed the float error

+++ train_new.py
@@ -50,8 +50,8 @@ def train_labeled_and_unlabeled(train_loader_with_label, train_loader_without_la
                 train_iter_without_label = iter(train_loader_without_label)
                 images_without_label, _ = next(train_iter_without_label)
 
-            images_with_label, labels = images_with_label.to(device), labels.to(device)
-            images_without_label = images_without_label.to(device)
+            images_with_label, labels = images_with_label.to(device, dtype=torch.float32), labels.to(device, dtype=torch.float32)
+            images_without_label = images_without_label.to(device, dtype=torch.float32)
             
             # zero the parameter gradients
             optimizer.zero_grad()
@@ -88,7 +88,7 @@ def train_labeled_and_unlabeled(train_loader_with_label, train_loader_without_la
         
         with torch.no_grad():
             for data, target in val_loader:
-                data, target = data.to(device), target.to(device)
+                data, target = data.to(device, dtype=torch.float32), target.to(device, dtype=torch.float32)
                 output = model(data)
                 
                 loss = criterion(output, target)
